requirements.txt
streamlit>=1.35.0
anthropic>=0.28.0
httpx>=0.27.0
dnspython>=2.6.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
cryptography>=42.0.0
